backend/requirements.txt
# EcoDrive Simulator Requirements
# For production deployment, use: requirements-prod.txt
# This file includes all dependencies for development and production

# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Environment configuration
python-dotenv==1.0.0

# Production WSGI server (needed for Render, Heroku, etc.)
# Note: Optional for local development, required for production
gunicorn==21.2.0
gevent==23.9.1

# ML/Data Science dependencies
numpy>=1.25.0
pandas>=2.0.3
scikit-learn>=1.3.0
matplotlib>=3.8.3
seaborn>=0.12.2
stable-baselines3>=2.3.0
gymnasium>=0.29.1,<0.30.0
torch>=2.6.0

# F1Tenth dependencies
# Note: f1tenth_gym must be installed from local f1tenth_rl-main directory
# See installation instructions in README.md
opencv-python>=4.9.0.80
Pillow>=10.0.1
numba>=0.58.0
scipy>=1.11.0

# FastF1 API for F1 data
fastf1

# System monitoring
psutil>=5.9.0

# Utilities
fire>=0.5.0
